Re-distrubted access.py functions to separate classes
