+++ src/transactions/transaction.py
@@ -6,9 +6,6 @@ verifying
 """
 from base64 import b64encode, b64decode
 from hashlib import sha256
-from ecdsa import VerifyingKey, NIST256p
-from ecdsa import BadSignatureError
-from src.persistence import access
 
 
 class Transaction(object):
@@ -164,79 +161,6 @@ class Transaction(object):
                 sender_public_key.to_string()
                 ).decode()
         return self.transaction_id
-
-    @staticmethod
-    def verify(tnx):
-        """
-        Verify the authenticity of a transaction object.
-        Verifying checks the unlocking signature against the transaction,
-        which will only be authentic if nothing was tampered with. After this,
-        each input is checked against the blockchain for existence and right
-        of use by the sender.
-
-        Returns:
-            A tuple (boolean, dict) where the first is whether or not the
-            transaction is valid, and the second is None unless the transaction
-            was invalid by unrightful use of input, in which case it is the
-            offending input.
-        """
-        authentic = True
-        offender = None
-
-        sender_address = tnx.unlock['sender_public_key']  # sender's address
-        # The sender's public key was sent over as a string, we must
-        # convert it back into a Key object. First it is converted from its
-        # base64 encoding back into a byte string, and then from a byte string
-        # into a Key object
-        sender_byte_address = b64decode(sender_address.encode())
-        verify_key = VerifyingKey.from_string(sender_byte_address,
-                                              curve=NIST256p)
-        # The same must be done for the signature, which was also
-        # sent over as a string. COnvert from base64 encoding to byte string,
-        # and from byte string to Signature object
-        signature = b64decode(tnx.unlock['signature'].encode())
-        # Re-build the message from scratch to expose any
-        # invalid data in the transaction
-        message = tnx.get_message()
-
-        # The transaction's message (all protected data) is verified
-        # using the sender's public key. Because the sender signed with
-        # a private key and this public key is mathematically related to it,
-        # we can use it to authenticate the transaction signature.
-        try:
-            verify_key.verify(signature, message, hashfunc=sha256)
-        except BadSignatureError:
-            # An invalid signature can mean that the transaction was
-            # either tampered with, or the wrong public key was used.
-            authentic = False
-
-        if authentic:
-            # Check that the sender has the rights to use all of the
-            # inputs he uses in the transaction.
-            for t_input in tnx.inputs:
-                # find the refrenced output used as an input
-                refrenced_output = access.find_output(
-                    t_input['transaction_id'],
-                    t_input['block_id'],
-                    t_input['output_index'])
-
-                # make sure the refrenced output exists
-                if not refrenced_output:
-                    authentic = False
-                    offender = t_input
-                    break
-                # make sure refrenced output is addressed transaction's sender
-                elif refrenced_output['receiver_address'] != sender_address:
-                    authentic = False
-                    offender = t_input
-                    break
-                # make sure transaction output hasnt already been spent
-                elif refrenced_output['spent_transaction_id']:
-                    authentic = False
-                    offender = t_input
-                    break
-
-        return authentic, offender
 
     def get_message(self):
         """
